+++ Source/Regard.Common/API/Model/ApiVideo.cs
@@ -34,7 +34,7 @@ namespace Regard.Common.API.Model
         
         public string UploaderName { get; set; }
 
-        public int? Views { get; set; }
+        public ulong? Views { get; set; }
 
         public float? Rating { get; set; }
     }
